requirements.txt
altair==4.2.2
beautifulsoup4==4.11.2
certifi==2023.5.7
config==0.5.1
functions==0.7.0
ipython==8.13.2
Jinja2==3.1.2
lxml==4.9.1
matplotlib==3.7.1
numpy
pandas
pandas_datareader==0.10.0
pandas_ta==0.3.14b0
plotly==5.14.1
python_dateutil==2.8.2
pytz==2022.7.1
pygame==2.5.2
Requests==2.31.0
scipy==1.10.0
seaborn==0.12.2
streamlit==1.25.0
streamlit_on_Hover_tabs==1.0.1
ta==0.10.2
tqdm==4.64.0
tweepy==4.12.1
yahoo_fin==0.8.9.1
yfinance==0.2.18
nltk==3.7
xlsxwriter==3.0.3
pip==23.3.2
replicate
millify

